+++ CHANGELOG.md
@@ -20,6 +20,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   during building the segments, any segment that doesn't pass a
   defined verification is not added. Furthermore, such a segment's (or
   augment's) setup function will return early as well.
+- Segment `global-mode-string` is now `misc-info` and returns the
+  equivalent information in `mode-line-misc-info`.
+- Segment `process` now returns `mode-line-process` as-is.
 
 ### Removed